Fix crash when class declaration extends an expression.
Fixes #679.

It's possible for a class declaration to extend an expression that does
not have a symbol, for example when a mixin function is used to build a
base class, as in `declare MyClass extends MyMixin(MyBaseClass)`.

Handling this correctly is tricky. Closure throws on this
`extends <expression>` syntax (see
https://github.com/google/closure-compiler/issues/2182). We would
probably need to generate an intermediate class declaration and extend
that. For now, just omit the `extends` annotation.