Make set ts.setTextRange on side effect imports.
TypeScript does not respect the "NoComments" emit flag on import
statements. The workaround is to manually set the text range of the
node.

This already worked for regular imports and exports, but did not work
for side effect imports, which produce a different emit AST structure.
This fixes the detection logic to cover that case.